+++ kafka_influxdb/reader/kafka_python.py
@@ -1,12 +1,18 @@
 # -*- coding: utf-8 -*-
 
 import logging
-import time
 from kafka import KafkaConsumer
+from kafka.common import ConsumerTimeout
+
+from kafka_influxdb.encoder.errors import EncoderError
 
 
-class KafkaReader(object):
-    def __init__(self, host, port, group, topic, reconnect_wait_time=2):
+class Reader(object):
+    """
+    A Kafka consumer based on kafka-python
+    See: https://github.com/dpkp/kafka-python
+    """
+    def __init__(self, host, port, group, topic):
         """
         Initialize Kafka reader
         """
@@ -14,7 +20,6 @@ class KafkaReader(object):
         self.port = port
         self.group = group
         self.topic = topic
-        self.reconnect_wait_time = reconnect_wait_time
 
         # Initialized on read
         self.consumer = None
@@ -30,9 +35,8 @@ class KafkaReader(object):
         """
         Read from Kafka. Reconnect on error.
         """
-        while True:
-            for msg in self._handle_read():
-                yield msg
+        for msg in self._handle_read():
+            yield msg
 
     def _handle_read(self):
         """
@@ -42,8 +46,7 @@ class KafkaReader(object):
             self._connect()
             for message in self.consumer:
                 yield message.value
-        except Exception as e:
-            logging.error("Kafka error: %s.", e)
-            logging.error("Trying to reconnect to %s:%s", self.host, self.port)
-            time.sleep(self.reconnect_wait_time)
-            pass
+        except ConsumerTimeout as timeout:
+            logging.error("Kafka error: %s.", timeout)
+            # The actual reconnect handling is done in the worker
+            raise EncoderError(timeout.message)